+++ tests/FSharp.AWS.DynamoDB.Tests/ConditionalExpressionTests.fs
@@ -62,9 +62,6 @@ module CondExprTypes =
             Map : Map<string, int64>
 
             Set : Set<int64>
-
-            [<BinaryFormatter>]
-            Serialized : int64 * string
         }
 
 type ``Conditional Expression Tests`` (fixture : TableFixture) =
@@ -85,7 +82,6 @@ type ``Conditional Expression Tests`` (fixture : TableFixture) =
             List = [for _ in 0L .. rand() % 5L -> rand() ]
             Array = [| for _ in 0L .. rand() % 5L -> rand() |]
             Union = if rand() % 2L = 0L then UA (rand()) else UB(guid())
-            Serialized = rand(), guid()
         }
 
     let table = fixture.CreateEmpty<CondExprRecord>()
@@ -103,6 +99,14 @@ type ``Conditional Expression Tests`` (fixture : TableFixture) =
         let _key = table.PutItem(item, precondition = itemDoesNotExist)
         fun () -> table.PutItem(item, precondition = itemDoesNotExist)
         |> shouldFailwith<_, ConditionalCheckFailedException>
+
+    let [<Fact(Skip = "Not sure if this should be working")>] ``Item not exists failure should add conflicting item data to the exception`` () =
+        let item = mkItem()
+        let _key = table.PutItem(item, precondition = itemDoesNotExist)
+        try
+            table.PutItem(item, precondition = itemDoesNotExist) |> ignore
+        with :? ConditionalCheckFailedException as e ->
+            test <@ e.Item.Count > 0  @>
 
     let [<Fact>] ``String precondition`` () =
         let item = mkItem()
@@ -376,13 +380,6 @@ type ``Conditional Expression Tests`` (fixture : TableFixture) =
         fun () -> table.Template.PrecomputeConditionalExpr <@ fun r -> r.Bytes.Length = r.Bytes.Length @>
         |> shouldFailwith<_, ArgumentException>
 
-
-    let [<Fact>] ``Serializable precondition`` () =
-        let item = mkItem()
-        let _key = table.PutItem item
-        fun () -> table.PutItem(item, <@ fun r -> r.Serialized = (0L,"")  @>)
-        |> shouldFailwith<_, ArgumentException>
-
     let [<Fact>] ``EXISTS precondition`` () =
         let item = { mkItem() with List = [1L] }
         let _key = table.PutItem item
@@ -466,7 +463,6 @@ type ``Conditional Expression Tests`` (fixture : TableFixture) =
         test false <@ fun r -> r.Map > Map.empty @>
         test false <@ fun r -> r.Set > Set.empty @>
         test false <@ fun r -> r.Ref > ref "12" @>
-        test false <@ fun r -> r.Serialized <= (1L, "32") @>
         test false <@ fun r -> r.Tuple <= (1L, 2L) @>
         test false <@ fun r -> r.Nested <= r.Nested @>